🚚 rename: InformationCard 컴포넌트를 MyTicket 내부 components 의 폴더로 이동 #46
